requirements.txt
numpy
wheel
python-docx
pymupdf
easyocr
langdetect
rapidfuzz
deep-translator
spacy
fastapi[standard]
ipython
annotated-types==0.7.0
asttokens==3.0.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
comm==0.2.2
confection==0.1.5
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
langcodes==3.5.0
language-data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.13
nest-asyncio==1.6.0
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.55
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
preshed==3.0.10
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
pymupdf==1.26.0
python-dateutil==2.9.0.post0
pyyaml==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
thinc==8.3.6
tldextract==5.3.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
PyMuPDF
python-docx
easyocr
pywin32
numpy
opencv-python
pillow
langdetect
pywin32
deep_translator
presidio-analyzer
presidio-anonymizer
unidecode